+++ terra_utils/urls.py
@@ -1,9 +1,0 @@
-from django.urls import path
-
-from terra_utils.views import SettingsView
-
-app_name = 'terra_utils'
-
-urlpatterns = [
-    path('settings/', SettingsView.as_view(), name='settings'),
-]